# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store new user",
                "description": "Returns users data",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login",
                "description": "Returns users data",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "getPostList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page to filter by",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "PostResource": {
                "title": "PostResource",
                "description": "Post resource",
                "xml": {
                    "name": "PostResource"
                }
            },
            "Author": {
                "title": "Author",
                "description": "Author model",
                "xml": {
                    "name": "Author"
                }
            },
            "Post": {
                "title": "Post",
                "description": "Post model",
                "xml": {
                    "name": "Post"
                }
            },
            "Subscription": {
                "title": "Subscription",
                "description": "Subscription model",
                "xml": {
                    "name": "Subscription"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "xml": {
                    "name": "User"
                }
            },
            "AuthUserRequest": {
                "title": "Store User request",
                "description": "Store User request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email of user",
                        "type": "string",
                        "example": "john007@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of user",
                        "type": "string",
                        "example": "2sdd243523fdg35"
                    }
                },
                "type": "object"
            },
            "StoreUserRequest": {
                "title": "Store User request",
                "description": "Store User request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of a user",
                        "type": "string",
                        "example": "John"
                    },
                    "email": {
                        "title": "email",
                        "description": "Email of a user",
                        "type": "string",
                        "example": "john007@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of a user",
                        "type": "string",
                        "example": "2sdd243523fdg35"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "Password of a user to confirm",
                        "type": "string",
                        "example": "2sdd243523fdg35"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Posts",
            "description": "API Endpoints of Posts"
        }
    ],
    "security": [
        []
    ]
}